Refactor the Vuex store.
