+++ build_scripts/dev-client.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 require('eventsource-polyfill')
-var hotClient = require('webpack-hot-middleware/client?noInfo=true&reload=true')
+var hotClient = require('webpack-hot-middleware/client?noInfo=true&reload=true&path=http://localhost:8080/__webpack_hmr&timeout=20000')
 
 hotClient.subscribe(function (event) {
   if (event.action === 'reload') {
